pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trusteval-ai"
version = "1.0.3"
description = "Enterprise LLM Evaluation & Responsible AI Framework for Healthcare, BFSI, Retail & Legal industries"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
  { name = "Antrixsh Gupta", email = "antrixsh@example.com" }
]
maintainers = [
  { name = "Antrixsh Gupta", email = "antrixsh@example.com" }
]
keywords = [
  "llm", "evaluation", "responsible-ai", "healthcare-ai", "bfsi",
  "bias-detection", "hallucination", "pii", "toxicity",
  "openai", "anthropic", "gemini", "huggingface",
  "enterprise-ai", "ai-safety", "llm-benchmark"
]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Software Development :: Quality Assurance",
]
requires-python = ">=3.10"

dependencies = [
  "openai>=1.0.0",
  "anthropic>=0.20.0",
  "google-generativeai>=0.4.0",
  "transformers>=4.38.0",
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
  "sqlalchemy>=2.0.0",
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
  "click>=8.1.0",
  "rich>=13.0.0",
  "cryptography>=42.0.0",
  "python-jose>=3.3.0",
  "slowapi>=0.1.9",
  "loguru>=0.7.0",
  "weasyprint>=61.0",
  "pandas>=2.0.0",
  "numpy>=1.24.0",
  "scikit-learn>=1.3.0",
  "httpx>=0.27.0",
  "tenacity>=8.2.0",
  "python-dotenv>=1.0.0",
  "aiosqlite>=0.20.0",
  "websockets>=12.0",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-mock>=3.12.0",
  "ruff>=0.3.0",
  "mypy>=1.9.0",
  "bandit>=1.7.0",
  "safety>=3.0.0",
  "responses>=0.25.0",
]

[project.scripts]
trusteval = "cli.main:cli"

[project.urls]
Homepage = "https://github.com/antrixsh/trusteval"
Documentation = "https://github.com/antrixsh/trusteval/docs"
Repository = "https://github.com/antrixsh/trusteval"
"Bug Tracker" = "https://github.com/antrixsh/trusteval/issues"
"Author LinkedIn" = "https://www.linkedin.com/in/antrixshgupta"

[tool.setuptools.packages.find]
include = ["trusteval*", "cli*", "dashboard*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
